désactivation de filtrage de photos avec lez service ImageManipulator

+++ src/Controller/Admin/ProductsController.php
@@ -25,7 +25,7 @@ class ProductsController extends AbstractController
         return $this->render('admin/products/index.html.twig',compact('products'));
     }
     #[Route('/ajout', name: 'add')]
-    public function add(Request $request , EntityManagerInterface $em , SluggerInterface $slugger):Response
+    public function add(ImageManipulator $imageManipulator,Request $request , EntityManagerInterface $em , SluggerInterface $slugger):Response
     {
         $this->denyAccessUnlessGranted('ROLE_ADMIN');
         //on creer un nouveau produit
@@ -35,23 +35,18 @@ class ProductsController extends AbstractController
        
         if($form->isSubmitted() && $form->isValid())
         {
-            //on recupere les images
-            // $images = $form->get('images')->getData();
-            // foreach($images as $image)
-            // {
-            //     //on definit le dossier de destination
-            //     $folder = 'products';
+        //     $imageFile  = $form->get('imageFile')->getData();
+        //     if($imageFile){
+        //     $targetDir = '/home/yazid-carpro/projets/kingdom-auto-care/public/assets/images/products/';
+        //     $imageManipulator->resize($product,$imageFile, 300, 300, $targetDir);
 
-            //     //on apl le service pictureService pour ajouter l'image
-            //     $fichier  =$pictureService->add($image , $folder , 300 , 300);
-            //     $img = new Images();
-            //     $img->setName($fichier);
-            //     $product->addImage($img);
-        
-            // }
+        //     $imageName = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME). '.webp';
+            
+        //     $product->setImage($imageName);
+        // }
+            
             $slug = $slugger->slug($product->getName());
             $product->setSlug($slug);
-            dd($product->getImage());
             $em->persist($product);
             $em->flush(); 
             $this->addFlash('success','Le produit a bien été ajouté');
@@ -77,20 +72,6 @@ class ProductsController extends AbstractController
         if($form->isSubmitted() && $form->isValid())
         {
 
-            //on recupere les images
-            $images = $form->get('images')->getData();
-            foreach($images as $image)
-            {
-                //on definit le dossier de destination
-                $folder = 'products';
-
-                //on apl le service pictureService pour ajouter l'image
-                $fichier  =$pictureService->add($image , $folder , 300 , 300);
-                $img = new Images();
-                $img->setName($fichier);
-                $product->setImage($img);
-        
-            }
             $slug = $slugger->slug($product->getName());
             $product->setSlug($slug);
             
@@ -100,40 +81,53 @@ class ProductsController extends AbstractController
         }
         return $this->render('admin/products/edit.html.twig' , ['form'=>$form->createView(),'product'=>$product]);
     }
+
+
+
+
     #[Route('/suppression/{id}', name: 'delete')]
-    public function delete(Products $product , EntityManagerInterface $em):Response
+    public function delete($id, EntityManagerInterface $em, ProductsRepository $repository): Response
     {
-    $this->denyAccessUnlessGranted('PRODUCT_DELETE', $product);
+        $product = $repository->find($id);
+    
+        if (!$product) {
+            $this->addFlash('error', 'Le produit demandé n\'existe pas.');
+            return $this->redirectToRoute('admin_products_index');
+        }
+    
+        $this->denyAccessUnlessGranted('PRODUCT_DELETE', $product);
+    
         $em->remove($product);
         $em->flush();
-        $this->addFlash('success','Le produit a bien été supprimé');
-        return $this->render('admin/products/index.html.twig');
+    
+        $this->addFlash('success', 'Le produit a bien été supprimé');
+        return $this->redirectToRoute('admin_products_index');
     }
 
-    //Delete image method
-    #[Route('/suppression/image/{id}', name: 'delete_image', methods: ['DELETE'])]
-    public function deleteImage(Images $image, Request $request, EntityManagerInterface $em, PictureService $pictureService): JsonResponse
-    {
-        // On récupère le contenu de la requête
-        $data = json_decode($request->getContent(), true);
+    // //Delete image method
+    // #[Route('/suppression/image/{id}', name: 'delete_image', methods: ['DELETE'])]
+    // public function deleteImage(Images $image, Request $request, EntityManagerInterface $em, PictureService $pictureService): JsonResponse
+    // {
+    //     // On récupère le contenu de la requête
+    //     $data = json_decode($request->getContent(), true);
 
-        if($this->isCsrfTokenValid('delete' . $image->getId(), $data['_token'])){
-            // Le token csrf est valide
-            // On récupère le nom de l'image
-            $nom = $image->getName();
+    //     if($this->isCsrfTokenValid('delete' . $image->getId(), $data['_token'])){
+    //         // Le token csrf est valide
+    //         // On récupère le nom de l'image
+    //         $nom = $image->getName();
 
-            if($pictureService->delete($nom, 'products', 300, 300)){
-                // On supprime l'image de la base de données
-                $em->remove($image);
-                $em->flush();
+    //         if($pictureService->delete($nom, 'products', 300, 300)){
+    //             // On supprime l'image de la base de données
+    //             $em->remove($image);
+    //             $em->flush();
 
-                return new JsonResponse(['success' => true], 200);
-            }
-            // La suppression a échoué
-            return new JsonResponse(['error' => 'Erreur de suppression'], 400);
-        }
+    //             return new JsonResponse(['success' => true], 200);
+    //         }
+    //         // La suppression a échoué
+    //         return new JsonResponse(['error' => 'Erreur de suppression'], 400);
+    //     }
 
-        return new JsonResponse(['error' => 'Token invalide'], 400);
-    }
+    //     return new JsonResponse(['error' => 'Token invalide'], 400);
+    // }
 
 }
